==> scripts/02_seamounts.Rmd <==
---
title: "Seamounts"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
---

```{r}
library(raster)
library(sf)
library(tidyverse)
library(ggspatial)
```

# Read spatial mgmt units

```{r}
load(file.path(project_path, "data", "02_processed_data", "admin","common_spatial_inputs.RData"))
```

# Seamounts datasets

```{r}
clark_seamounts <- readxl::read_xls(file.path(emlab_data_path, 
                                              "seamounts-clark", "Clarketal_2011_OCM_All_seamounts_values.xls"))

clark_seamounts <- clark_seamounts %>% 
  janitor::clean_names() %>% 
  select(id, long, lat, depth, class = smt_class_poc) %>% 
  mutate(depth_class = ifelse(depth <= 500, "<= 500m", "> 500m"))

col_smts_clark <- clark_seamounts %>% 
  st_as_sf(coords = c("long", "lat"),
           crs = 4326) %>% 
  st_transform(crs = col_crs) %>% 
  filter(st_within(., st_union(colombia_eez), sparse = FALSE)[,1])

yesson_smts <- sf::read_sf(file.path(emlab_data_path, "seamounts-yesson-2019", "YessonEtAl2019-Seamounts-V2.shp"))

col_smts_yesson <- yesson_smts %>% 
  janitor::clean_names() %>% 
  st_transform(crs = col_crs) %>% 
  filter(st_within(., st_union(colombia_eez), sparse = FALSE)[,1]) %>% 
  mutate(depth_class = ifelse(depth > -500, "<= 500m", "> 500m") )

col_smts <- col_smts_yesson %>%
  transmute(id = peak_id, depth = depth, depth_class = depth_class, source = "yesson") %>% 
  bind_rows(col_smts_clark %>% 
              transmute(id = id, depth = -depth, depth_class, source = "clark"))

col_background_plot+
  geom_sf(data = colombia_eez, linetype = "dashed", size = .25, fill = "transparent")+
  geom_sf(data = col_smts, 
          aes(col = fct_rev(depth_class)), size = .7, show.legend = T)+
  labs(alpha = "Profundidad", title = "Montes submarinos")+
  coord_sf(xlim = c(as.double(st_bbox(colombia_eez)["xmin"]), as.double(st_bbox(colombia_eez)["xmax"])+.1), 
           ylim = c(as.double(st_bbox(colombia_eez)["ymin"]), as.double(st_bbox(colombia_eez)["ymax"])), 
           expand = FALSE)+
  ggspatial::annotation_scale(location = "br", width_hint = 0.2, style = "ticks", 
                              pad_y = unit(0.05, units =  "native"), pad_x = unit(0.05, units =  "native") )+
  ggspatial::annotation_north_arrow(location = "tl", which_north = "true", height = unit(0.1, "native"), width = unit(0.1, "native"),
                                    pad_x = unit(0.05, "native"), pad_y = unit(0.05, "native"),
                                    style = north_arrow_fancy_orienteering)+
  paletteer::scale_color_paletteer_d(palette = "ochRe::lorikeet", name = "Profundidad")+
  ggsave(filename = file.path(project_path, "figures", "seamounts.png"), 
         dpi = 300)
```

# By ROI

## Offshore Pacific

```{r}
smts_leg_1 <- col_smts %>% 
  st_intersection(rois_offshore_pacific) 

smts_leg_1 %>% 
  group_by(region) %>% 
  summarize(n = n(),
            n_sub_500 = sum(depth_class != "> 500m"))
```

```{r}
col_background_plot +
  geom_sf(data = col_smts %>% 
            filter(st_within(., st_union(colombia_pacific_eez), sparse = FALSE)[,1]),
         aes(col = fct_rev(depth_class)), size = .7)+
  geom_sf(data = colombia_pacific_eez, linetype = "dashed", size = .25, fill = "transparent")+
  geom_sf(data = yurupari, fill = "transparent", col = "orange")+
  geom_sf(data = malpelo, fill = "transparent", col = "lightblue")+
  geom_sf(data = rois_offshore_pacific, fill = "transparent", col = "navy")+
  labs(title = "Yurupari - Malpelo / Colinas y Lomas del Pacifico",
       subtitle = "Montes submarinos")+
  geom_sf(data = st_crop(col_land, colombia_pacific_eez))+
  coord_sf(xlim = c(as.double(st_bbox(colombia_pacific_eez)["xmin"]), as.double(st_bbox(colombia_pacific_eez)["xmax"])+.1), 
           ylim = c(as.double(st_bbox(colombia_pacific_eez)["ymin"]), as.double(st_bbox(colombia_pacific_eez)["ymax"])), 
           expand = FALSE)+
  ggspatial::annotation_scale(location = "br", width_hint = 0.2, style = "ticks", 
                              pad_y = unit(0.02, units =  "native"), pad_x = unit(0.02, units =  "native") )+
  ggspatial::annotation_north_arrow(location = "tl", which_north = "true", height = unit(0.1, "native"), width = unit(0.1, "native"),
                                    pad_x = unit(0.05, "native"), pad_y = unit(0.05, "native"),
                                    style = north_arrow_fancy_orienteering)+
  paletteer::scale_color_paletteer_d(palette = "ochRe::lorikeet", name = "Profundidad")+
  ggsave(filename = file.path(project_path, "figures", "seamounts_yurupari_malpelo.png"), 
         dpi = 300, width = 7, height = 5)
```
```{r}
smts_leg_1 %>% 
  sf::st_write(file.path(project_path, "data", "02_processed_data", "seamounts_leg_1.gpkg"))
```


## Regimen Comun

```{r}
smts_leg_3 <- col_smts %>% 
  st_intersection(col_jam_area) 

col_background_plot +
  geom_sf(data = col_smts %>% 
            filter(st_within(., st_union(col_jam_area), sparse = FALSE)[,1]),
          aes(col = fct_rev(depth_class)), size = .7)+
  geom_sf(data = col_jam_area, linetype = "dashed", size = .25, fill = "transparent")+
  geom_sf(data = bajos_eez, linetype = "dashed", size = .25, fill = "transparent")+
  labs(title = "Montes submarinos",subtitle = "Area de Regimen Comun Colombia - Jamaica")+
  geom_sf(data = st_crop(col_land, col_jam_area))+
  coord_sf(xlim = c(as.double(st_bbox(col_jam_area)["xmin"]), as.double(st_bbox(col_jam_area)["xmax"])+.1), 
           ylim = c(as.double(st_bbox(col_jam_area)["ymin"]), as.double(st_bbox(col_jam_area)["ymax"])), 
           expand = FALSE)+
  ggspatial::annotation_scale(location = "bl", width_hint = 0.2, style = "ticks", 
                              pad_y = unit(0.02, units =  "native"), pad_x = unit(0.02, units =  "native") )+
  ggspatial::annotation_north_arrow(location = "bl", which_north = "true", height = unit(0.1, "native"), width = unit(0.1, "native"),
                                    pad_x = unit(0.05, "native"), pad_y = unit(0.05, "native"),
                                    style = north_arrow_fancy_orienteering)+
  paletteer::scale_color_paletteer_d(palette = "ochRe::lorikeet", name = "Profundidad")+
  theme(plot.subtitle = element_text(margin = margin(t = 0, r = 0, b = .3, l = 0, unit = "pt")))+
      ggsave(filename = file.path(project_path, "figures", "seamounts_regimen_comun.png"), 
         dpi = 300)
```

```{r}
bind_rows(smts_leg_3, 
          smts_leg_1) %>% 
  replace_na(list(region = "Seaflower")) %>% 
  st_transform(crs = 4326) %>% 
  dplyr::mutate(lon = sf::st_coordinates(.)[,1],
                lat = sf::st_coordinates(.)[,2]) %>% 
  select(id, lon, lat, depth, roi = region) %>% 
  st_drop_geometry() %>% 
  mutate_at(vars(lon, lat), round, 5) %>% 
  write_csv(file.path(project_path, "data", "03_output_data", "seamounts_in_rois.csv"))
```

# INVEMAR drop stations in Colinas y Lomas

```{r}
invemar_drops <- read_csv(file.path(project_path, "data", "01_raw_data", "invemar_drop_coords.csv")) 

invemar_drops_sf <- invemar_drops %>% 
    st_as_sf(coords = c("longitud", "latitud"),
           crs = 4326) %>% 
  st_transform(crs = col_crs) 

invemar_drops_sf %>% 
  sf::st_write(file.path(project_path, "data", "02_processed_data", "invemar_drops.gpkg"))

```

```{r}
invemar_drops %>% 
  mutate(lon = measurements::conv_unit(longitud,
                                       to = 'deg_dec_min', 
                                       from = 'dec_deg'),
         lat = measurements::conv_unit(latitud,
                                       to = 'deg_dec_min', 
                                       from = 'dec_deg'))
```
