==> scripts/04_rocky_reefs.Rmd <==
---
title: "04_rocky_reefs"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
---

```{r}
library(raster)
library(sf)
library(tidyverse)
library(ggspatial)
```

```{r}
load(file.path(project_path, "data", "02_processed_data", "admin","common_spatial_inputs.RData"))
```

# Chasqui et al

```{r}
riscales_raw <- readxl::read_xlsx(file.path(project_path, 
                                            "data", "01_raw_data", "riscales_chasqui.xlsx"))
```

```{r}
riscales_chasqui_clean <- riscales_raw %>% 
  janitor::clean_names() %>% 
  mutate(source = "chasqui") %>% 
  select(source, name = estaciones,  lon_dec = longitude_dec, lat_dec = latitude_dec) %>% 
  mutate(lon = measurements::conv_unit(lon_dec,
                                       from = "dec_deg",
                                       to = "deg_dec_min"),
         lat = measurements::conv_unit(lat_dec,
                                       from = "dec_deg",
                                       to = "deg_dec_min")) %>% 
  mutate(lon = stringr::str_trunc(lon, 11, "right", ellipsis = ""),
         lat = stringr::str_trunc(lat, 11, "right", ellipsis = ""))
```

```{r}
chasqui_sf <- riscales_chasqui_clean %>% 
  sf::st_as_sf(coords = c("lon_dec", "lat_dec"), 
               crs = 4326) %>% 
  st_transform(crs = col_crs) 
```


```{r}
chasqui_sf <- chasqui_sf %>% 
  st_join(bind_rows(drmi_gtcc["nombre"], 
                    utria["nombre"], 
                    zepa["nombre"],
                    zemp["nombre"]) %>% 
            rename(roi = nombre),
          left = T) 

chasqui_sf <- chasqui_sf %>% 
  replace_na(list(roi = "golfo de tribuga cabo corrientes")) 

chasqui_sf <- chasqui_sf %>% 
  cbind(st_coordinates(.)) %>%
  arrange(desc(Y)) %>%
  select(-X, -Y) %>% 
  mutate(roi = if_else(row_number() <= 7, "cabo marzo", roi))
```

```{r}
chasqui_sf %>% 
  group_by(roi) %>% 
  summarize(n_riscales = n())
```

```{r}
chasqui_sf %>% 
  sf::st_write(file.path(project_path, "data", "02_processed_data", "riscales_chasqui.gpkg"))
```

# Riscales TEK

```{r}
riscales_tek <- read_csv(file = file.path(project_path, 
                          "data", "01_raw_data","riscales_tek.csv")) %>% 
  arrange(desc(lat)) %>% 
  select(source, name, lon, lat, depth_m, depth_b) 

riscales_tek <- riscales_tek %>% 
  mutate(lon = paste0("-", lon),
         lon_dec = as.numeric(measurements::conv_unit(lon,
                                           from = "deg_dec_min",
                                           to = "dec_deg")),
         lat_dec = as.numeric(measurements::conv_unit(lat,
                                           from = "deg_dec_min",
                                           to = "dec_deg"))) %>% 
  mutate_if(is.numeric,round, 4) %>% 
  select(source, name, lon, lat, lon_dec, lat_dec, depth_m, depth_b)

riscales_tek %>% 
  sf::st_as_sf(coords = c("lon_dec", "lat_dec"), 
               crs = 4326) %>% 
  st_transform(crs = col_crs) %>% 
  st_write(file.path(project_path, 
                     "data", "03_output_data", "riscales_tek.gpkg"), append = F)
```

# Puntos Sandra

```{r}
pts_sandra <- read_csv(file = file.path(project_path, 
                          "data", "01_raw_data","puntos_golfo_sandra.csv")) %>% 
  arrange(desc(lat))

pts_sandra <- pts_sandra %>% 
  mutate(lon = paste0("-", lon),
         lon_dec = as.numeric(measurements::conv_unit(lon,
                                           from = "deg_min_sec",
                                           to = "dec_deg")),
         lat_dec = as.numeric(measurements::conv_unit(lat,
                                           from = "deg_min_sec",
                                           to = "dec_deg"))) %>% 
  mutate_if(is.numeric,round, 4) %>% 
  select(source, name = nombre, lon, lat, lon_dec, lat_dec)

pts_sandra <- pts_sandra %>% 
  mutate(lon = measurements::conv_unit(lon,
                                       from = "deg_min_sec",
                                       to = "deg_dec_min"),
         lat = measurements::conv_unit(lat,
                                       from = "deg_min_sec",
                                       to = "deg_dec_min")) %>% 
  mutate(lon = stringr::str_trunc(lon, 8, "right", ellipsis = ""),
         lat = stringr::str_trunc(lat, 8, "right", ellipsis = ""))

pts_sandra %>% 
  sf::st_as_sf(coords = c("lon_dec", "lat_dec"), 
               crs = 4326) %>% 
  st_transform(crs = col_crs) %>% 
  st_write(file.path(project_path, 
                     "data", "03_output_data", "pts_sandra.gpkg"), append = F)
```
# Turqui

```{r}
pts_turqui <- read_csv(file = file.path(project_path, "data", "01_raw_data","turqui_pts.csv")) 

pts_turqui <- pts_turqui %>% 
  rename(depth_m = depth) %>% 
  mutate(source = "turqui",
         lon_dec = as.numeric(measurements::conv_unit(lon,
                                                      from = "deg_dec_min",
                                                      to = "dec_deg")),
         lat_dec = as.numeric(measurements::conv_unit(lat,
                                                      from = "deg_dec_min",
                                                      to = "dec_deg")))

pts_turqui %>% 
  sf::st_as_sf(coords = c("lon_dec", "lat_dec"), 
               crs = 4326) %>% 
  st_transform(crs = col_crs) %>% 
  st_write(file.path(project_path, 
                     "data", "03_output_data", "pts_turqui.gpkg"), append = F)
```

# Combine them

```{r}
riscales_coords <- pts_sandra %>% 
  bind_rows(riscales_tek) %>% 
  bind_rows(riscales_chasqui_clean) %>% 
  bind_rows(pts_turqui) %>% 
  arrange((lat_dec))

riscales_coords %>% 
  replace_na(list("depth_m" = "", "depth_b" = "")) %>% 
  write_csv(file.path(project_path, "data", "03_output_data", "riscales_coords.csv"))
```

```{r}
rocky_reefs <- sf::read_sf(file.path(project_path, "/data/01_raw_data/Objetos\ de\ Conservación\ PNCh/FRocosos/Arrecifesrocosos.shp")) %>% 
  st_transform(col_crs)
```

```{r}
col_background_plot+
  geom_sf(data = st_crop(col_land, uac_pnch), size = 0.2)+
    geom_sf(data = rocky_reefs, 
            fill = "#875632",
          col = "transparent")+
  geom_sf(data = filter(riscales_clean_sf, roi != "Zona Exclusiva de Pesca Artesanal"), 
          size = .7,
          shape = 25, col = "black")+
  coord_sf(xlim = c(as.double(st_bbox(uac_pnch)["xmin"]), as.double(st_bbox(uac_pnch)["xmax"])+.1), 
           ylim = c(as.double(st_bbox(uac_pnch)["ymin"]), as.double(st_bbox(uac_pnch)["ymax"])), 
           expand = FALSE)+
  ggspatial::annotation_scale(location = "br", width_hint = 0.2, style = "ticks", 
                              pad_y = unit(0.02, units =  "native"), pad_x = unit(0.02, units =  "native") )+
  ggspatial::annotation_north_arrow(location = "tr", which_north = "true", height = unit(0.1, "native"), width = unit(0.1, "native"),
                                    pad_x = unit(0.05, "native"), pad_y = unit(0.05, "native"),
                                    style = north_arrow_fancy_orienteering)+
  labs(title = "Arrecifes Rocosos", col = "")+
  paletteer::scale_color_paletteer_d("ggsci::default_jama", 
                                     labels = c("Cabo Marzo", "DRMI-GTCC", "PNN Utria", "ZEPA"))+
  geom_point(data = riscales_clean_sf %>% 
               filter(nombre %in% c("La Mina", "Amargal")),
             aes( geometry = geometry),
             stat = "sf_coordinates", 
             shape = 8,
             col = "lightblue")+
    ggsave(filename = file.path(project_path, "figures", "rocky_reefs.png"), 
         dpi = 300)
```
```{r}
riscales_clean_sf %>% 
  filter(roi != "Zona Exclusiva de Pesca Artesanal" | nombre == "La Mina") %>% 
  st_transform(crs = 4326) %>% 
  dplyr::mutate(lon = sf::st_coordinates(.)[,1],
                lat = sf::st_coordinates(.)[,2]) %>% 
  select(nombre, lon, lat, roi, mesophotic) %>% 
  st_drop_geometry() %>% 
  mutate_at(vars(lon, lat), round, 5) %>% 
  write_csv(file.path(project_path, "data", "03_output_data", "riscales_in_rois.csv"))
```

## mapa con mangalres

```{r}
gtcc_roi_bbox <- st_union(drmi_gtcc, utria) %>% 
  st_buffer(dist = 10000) %>% 
  st_difference(col_land) %>% 
  st_bbox()

gtcc_roi_bbox["xmin"]
```

```{r}
mangroves_uac_pnch <- read_sf(file.path(project_path, "data", "02_processed_data", "mangroves_uac_pnch.gpkg")) %>% 
  st_transform(crs = col_crs)

col_background_plot+
  geom_sf(data = st_crop(col_land, 
                         st_union(drmi_gtcc, utria) %>% 
                           st_buffer(dist = 10000) ))+
  geom_sf(data = drmi_gtcc, fill = "transparent", linetype ="dashed")+
  geom_sf(data = utria, fill = "transparent", linetype ="dashed")+
  geom_sf(data = st_crop(mangroves_uac_pnch, st_union(drmi_gtcc, utria)), 
          col = "transparent", fill = "darkgreen")+
  geom_sf(data = rocky_reefs, 
          fill = "#875632",
          col = "transparent")+
  geom_sf(data = filter(riscales_clean_sf, roi != "Zona Exclusiva de Pesca Artesanal"), 
          size = .7,
          shape = 25, col = "yellow", fill = "transparent")+
  coord_sf(xlim = c(gtcc_roi_bbox["xmin"], gtcc_roi_bbox["xmax"] + 4000), 
           ylim = c(gtcc_roi_bbox["ymin"] + 5000, gtcc_roi_bbox["ymax"]), 
           expand = FALSE)+
  ggspatial::annotation_scale(location = "br", width_hint = 0.2, style = "ticks", 
                              pad_y = unit(0.05, units =  "native"), pad_x = unit(0.05, units =  "native") )+
  ggspatial::annotation_north_arrow(location = "tl", which_north = "true", height = unit(0.1, "native"), width = unit(0.1, "native"),
                                    pad_x = unit(0.05, "native"), pad_y = unit(0.05, "native"),
                                    style = north_arrow_fancy_orienteering)+
  labs(title = "Manglares y arrecifes rocosos", subtitle = "Golfo de Tribugá - Cabo Corrientes")+
  ggsave(filename = file.path(project_path, "figures", "roi_gtcc.png"), 
         dpi = 300)

```
## INVEMAR/ANH stations

```{r}
invemar_stations <- read_csv(file.path(project_path, "data", "01_raw_data","estaciones_AUNAP_INVEMAR.csv"))

# change the degree symbol to a space
invemar_stations$lat = gsub('°', ' ', invemar_stations$lat)
invemar_stations$lat = gsub(',', '.', invemar_stations$lat)

invemar_stations$lon = gsub('°', ' ', invemar_stations$lon)
invemar_stations$lon = gsub(',', '.', invemar_stations$lon)


# convert from decimal minutes to decimal degrees
invemar_stations$lat_d = as.numeric(measurements::conv_unit(invemar_stations$lat, from = 'deg_dec_min', to = 'dec_deg'))
invemar_stations$lon_d = as.numeric(measurements::conv_unit(invemar_stations$lon, from = 'deg_dec_min', to = 'dec_deg'))

invemar_stations %>% 
  select(aunap_station_id, station_name, lat_d, lon_d) %>% 
  mutate(lon_d = -1*lon_d) %>% 
  sf::st_as_sf(coords = c("lon_d", "lat_d"), crs = 4326) %>% 
  st_transform(crs = st_crs(bajos_eez)) %>% 
  st_write(file.path(project_path, "data", "02_processed_data", "invemar_anh_stations.gpkg"), append = F)

```

```{r}
-77.43400,5.65561

measurements::conv_unit(15.8718, from = 'dec_deg', to = 'deg_dec_min')
```

```{r}
measurements::conv_unit("77 24.834",  from = 'deg_dec_min', to = 'dec_deg')
```
```{r}
measurements::conv_unit("15 48.195",  from = 'deg_dec_min', to = 'dec_deg')

measurements::conv_unit("78 45.979",  from = 'deg_dec_min', to = 'dec_deg')
```


## Riscales TEK

```{r}
read_csv(file = file.path(project_path, "data", "03_output_data","riscales_in_rois.csv")) %>% 
  filter(source == "tek")


  arrange(desc(lat)) %>% 
  rowid_to_column() %>% 
  select(id = rowid, source, nombre, lon, lat, profundidad_m) %>% 
  sf::st_as_sf(coords = c("lon", "lat"), 
               crs = 4326) %>% 
  st_transform(crs = col_crs) %>% 
  st_write(file.path(project_path, "data", "03_output_data", "riscales.gpkg"), append = F)
```

## CC point

```{r}
tibble("lat" = 217685.506,
        "lon" = 620530.965) %>% 
  sf::st_as_sf(coords = c("lon", "lat"), 
               crs = col_crs) %>% 
  st_transform(crs = 4326) %>% 
  dplyr::mutate(lon = sf::st_coordinates(.)[,1],
                lat = sf::st_coordinates(.)[,2]) %>% 
  ggplot()+
  geom_point(aes(lon, lat))+
  geom_sf(data = drmi_gtcc, fill = "transparent", linetype ="dashed")
```


```{r}
slopes <- raster("/Volumes/juan ex ssd/COL_bathy_data/sampling grid/Slope/slopetribu100.ovr", 
                 crs = "+proj=utm +zone=18 +datum=WGS84 +units=m +no_defs")

slopes %>% 
  projectRaster(tribuga_bati)
```


```{r}
measurements::conv_unit(x = 80, from = "ft", "m" )
```
