==> scripts/13_process_fish.Rmd <==
---
title: "Process fish data"
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: show
    toc: true
    toc_depth: 1
    toc_float: yes
    number_sections: false
    theme: cerulean
    highlight: haddock
---

```{r}
library(tidyverse)
library(ggspatial)
library(raster)
library(sf)
library(ggspatial)
```

# Read metadata

```{r}
load(file.path(project_path, "data", "02_processed_data", "admin","common_spatial_inputs.RData"))

fish_metadata <- readxl::read_xlsx("/Volumes/ps-col-exp/data/01_raw/fish/COL_fish_amf.xlsx", 
                                   sheet = "fish_metadata")
```

```{r}
gtcc_roi_bbox <- st_union(drmi_gtcc, utria) %>% 
  st_buffer(dist = 10000) %>% 
  st_difference(col_land) %>% 
  st_bbox()
```

```{r}
gtcc_background_plot <- col_background_plot +
  geom_sf(data = st_crop(col_land, 
                         st_union(drmi_gtcc, utria) %>% 
                           st_buffer(dist = 10000)))+
  geom_sf(data = drmi_gtcc, fill = "transparent", linetype ="dashed")+
  geom_sf(data = utria, fill = "transparent", linetype ="dashed")+
  coord_sf(xlim = c(gtcc_roi_bbox["xmin"], gtcc_roi_bbox["xmax"] + 4000), 
           ylim = c(gtcc_roi_bbox["ymin"] + 5000, gtcc_roi_bbox["ymax"]), 
           expand = FALSE)+
  ggspatial::annotation_scale(location = "br", width_hint = 0.2, style = "ticks", 
                              pad_y = unit(0.05, units =  "native"), 
                              pad_x = unit(0.05, units =  "native") )+
  ggspatial::annotation_north_arrow(location = "tl", which_north = "true", 
                                    height = unit(0.1, "native"), 
                                    width = unit(0.1, "native"),
                                    pad_x = unit(0.05, "native"), 
                                    pad_y = unit(0.05, "native"),
                                    style = north_arrow_fancy_orienteering)
```

```{r}
fish_meta_sf <- fish_metadata %>% 
  sf::st_as_sf(coords = c("lon", "lat"), 
               crs = 4326) %>% 
  st_transform(crs = col_crs) %>% 
  st_crop(gtcc_roi_bbox)
```

```{r}
gtcc_background_plot +
  geom_sf(data = fish_meta_sf)+
  coord_sf(xlim = c(gtcc_roi_bbox["xmin"], gtcc_roi_bbox["xmax"] + 4000), 
           ylim = c(gtcc_roi_bbox["ymin"] + 5000, gtcc_roi_bbox["ymax"]), 
           expand = FALSE)
```

```{r}
fish_metadata %>% 
  mutate(sub_location = location, 
         location = "Tribugá") %>% 
  distinct(ps_station_id, date, location, sub_location, lat, lon) %>% 
  write_csv("/Volumes/ps-col-exp/data/02_processed/fish_metadata.csv")
```
